# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Happy App Backend Api",
        "contact": {
            "name": "Mahbubul Alam",
            "email": "mahbubulalam5676@gmail.com"
        },
        "version": "1"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "User RestApi"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "int",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mahbubul@orangetoolz.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Correct credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payload": {
                                            "type": "collection",
                                            "example": {
                                                "user": {
                                                    "id": 1,
                                                    "company_id": 1,
                                                    "is_parent": 1,
                                                    "name": "Mahbubul Alam",
                                                    "email": "mahbubul@orangetoolz.com",
                                                    "email_verified_at": "2021-07-04T09:28:02.000000Z",
                                                    "phone_number": "01906043504",
                                                    "profile_image": null,
                                                    "email_signature": null,
                                                    "logged_at": null,
                                                    "status": 1,
                                                    "created_at": "2021-07-04T09:28:02.000000Z",
                                                    "updated_at": "2021-07-04T09:28:02.000000Z"
                                                },
                                                "token": {
                                                    "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9hcGkucmFsbGUudGVzdFwvYXBpXC92MVwvbG9naW4iLCJpYXQiOjE2MjQ2MzA5MDcsImV4cCI6MTYyNDYzNDUwNywibmJmIjoxNjI0NjMwOTA3LCJqdGkiOiJ6cE80VDdBVUJRbjMyRkU5Iiwic3ViIjoxLCJwcnYiOiJiOTEyNzk5NzhmMTFhYTdiYzU2NzA0ODdmZmYwMWUyMjgyNTNmZTQ4In0.egbMdLZlfz_WKyQj7rPsqn6Y0ynxHWKvhdfDABD1gXQ",
                                                    "token_type": "bearer",
                                                    "expires_in": 3600
                                                }
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged in"
                                        },
                                        "status": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "status_code": {
                                            "type": "int",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payload": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        },
                                        "status": {
                                            "type": "bool",
                                            "example": "false"
                                        },
                                        "status_code": {
                                            "type": "int",
                                            "example": "401"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payload": {
                                            "type": "collection",
                                            "example": {
                                                "email": [
                                                    "The email must be at least 8 characters.",
                                                    "The email must be a valid email address."
                                                ],
                                                "password": [
                                                    "The password must be at least 8 characters."
                                                ]
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Given data is invalid"
                                        },
                                        "status": {
                                            "type": "bool",
                                            "example": "false"
                                        },
                                        "status_code": {
                                            "type": "int",
                                            "example": "422"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}